refactor(core): split analysis/intraday.ts into intraday/ modules
Break the 1092-line orchestrator into cohesive files under
packages/core/src/analysis/intraday/: constants (shared styling/lookup
tables), timeframe (per-timeframe MACD/divergence detection and
coercion), markers (overlay marker/connector builders), entryPlan
(entry-plan computation and price-zone normalization), and orchestrator
(buildIntraday and its input/meta types). intraday.ts now re-exports
the same public API from its original path so every existing import
site keeps working unchanged.
